requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
razorpay==1.4.2
reportlab==4.3.1
requests==2.32.3
setuptools==75.8.2
sqlparse==0.5.3
urllib3==2.3.0
